[1689-verses-normalization] Normalizing verses in the 1689 pt 1:
🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥
Not only is this confession the only one with heterodox
views on baptism...
🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥

It is the only one with the value of an array for the verses property.

Before we standardize this data structure as it is on other
confessions, we are eliminating this case where
the member in the array represents mutliple scripture
references w/o specifying a range.

This makes our verses data more consistent for the time being. 👌👌👌👌

See issue #14 for further context.

(reformed-standards/compendium#14)