Refactor Validate Verses Test
Prior to this commit, we were just providing article[number]
as the location of the validation failure. This was not
adequate for documents where the failing citation is on an
article whose number is relative to the chapter rather than
the document; for instance, savoy chapter 5 article 4 has a
failing citation. This was showing as a failure on "chapter 4"
but that was inaccurate -- it was article 4, under chapter 5.

Now, with this commit, we surface both the relevant chapter
and article. If no article is given, we default to 0.